Adds flushAsync and logKeepAlives files
