+++ packages/core/src/devices/temperatureControl.test.ts
@@ -6,7 +6,7 @@ const NAME = 'TemperatureControl';
 import { jest } from '@jest/globals';
 
 // Matterbridge
-import { loggerErrorSpy, loggerFatalSpy, loggerWarnSpy, setupTest } from '../jestutils/jestHelpers.js';
+import { loggerErrorSpy, loggerFatalSpy, loggerWarnSpy, setupTest } from '../jestutils/jestSetupTest.js';
 import { laundryDryer, laundryWasher } from '../matterbridgeDeviceTypes.js';
 import { MatterbridgeEndpoint } from '../matterbridgeEndpoint.js';
 import { createLevelTemperatureControlClusterServer, createNumberTemperatureControlClusterServer } from './temperatureControl.js';
