+++ packages/core/src/jestutils/jestHelpers.ts
@@ -61,51 +61,27 @@
  *  ```
  */
 
-import { rmSync } from 'node:fs';
 import path from 'node:path';
-import { inspect } from 'node:util';
 
-import type { jest } from '@jest/globals';
 // @matter
-import { Environment, Lifecycle, LogFormat as MatterLogFormat, LogLevel as MatterLogLevel } from '@matter/general';
-import { Endpoint, ServerNode, ServerNodeStore } from '@matter/node';
+import { Environment } from '@matter/general';
+import { Endpoint, ServerNode } from '@matter/node';
 import { AggregatorEndpoint } from '@matter/node/endpoints';
 import { MdnsService } from '@matter/protocol';
-import { DeviceTypeId, VendorId } from '@matter/types/datatype';
 // @matterbridge
 import { MATTER_STORAGE_DIR, NODE_STORAGE_DIR } from '@matterbridge/types';
-import { AnsiLogger, er, LogLevel, rs, TimestampFormat, UNDERLINE, UNDERLINEOFF } from 'node-ansi-logger';
+import { AnsiLogger, LogLevel, rs, TimestampFormat, UNDERLINE, UNDERLINEOFF } from 'node-ansi-logger';
 import { NodeStorageManager } from 'node-persist-manager';
 
 import type { DeviceManager } from '../deviceManager.js';
 import type { Frontend } from '../frontend.js';
 import { Matterbridge } from '../matterbridge.js';
-import { bridge } from '../matterbridgeDeviceTypes.js';
 import type { MatterbridgePlatform } from '../matterbridgePlatform.js';
 import type { PluginManager } from '../pluginManager.js';
+import { flushAsync } from './jestFlushAsync.js';
+import { assertAllEndpointNumbersPersisted, createTestEnvironment, flushAllEndpointNumberPersistence } from './jestMatterTest.js';
+import { HOMEDIR, loggerLogSpy, originalProcessArgv } from './jestSetupTest.js';
 
-// Freeze the original process arguments and environment variables to allow resetting them in tests
-export const originalProcessArgv = Object.freeze([...process.argv]);
-export const originalProcessEnv = Object.freeze({ ...process.env } as Record<string, string | undefined>);
-
-// Spy on logger methods
-export let loggerLogSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.log>;
-export let loggerDebugSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.debug>;
-export let loggerInfoSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.info>;
-export let loggerNoticeSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.notice>;
-export let loggerWarnSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.warn>;
-export let loggerErrorSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.error>;
-export let loggerFatalSpy: jest.SpiedFunction<typeof AnsiLogger.prototype.fatal>;
-
-// Spy on console methods
-export let consoleLogSpy: jest.SpiedFunction<typeof console.log>;
-export let consoleDebugSpy: jest.SpiedFunction<typeof console.debug>;
-export let consoleInfoSpy: jest.SpiedFunction<typeof console.info>;
-export let consoleWarnSpy: jest.SpiedFunction<typeof console.warn>;
-export let consoleErrorSpy: jest.SpiedFunction<typeof console.error>;
-
-export let NAME: string;
-export let HOMEDIR: string;
 export let matterbridge: Matterbridge;
 export let frontend: Frontend;
 export let plugins: PluginManager;
@@ -115,98 +91,6 @@ export let environment: Environment;
 export let server: ServerNode<ServerNode.RootEndpoint>;
 export let aggregator: Endpoint<AggregatorEndpoint>;
 export let log: AnsiLogger;
-
-/**
- * Setup the Jest environment:
- * - it will remove any existing home directory
- * - setup the spies for logging
- *
- * @param {string} name The name of the test suite.
- * @param {boolean} debug If true, the logging is not mocked.
- *
- * @example
- * ```typescript
- * import { consoleDebugSpy, consoleErrorSpy, consoleInfoSpy, consoleLogSpy, consoleWarnSpy, loggerLogSpy, setDebug, setupTest } from './jestutils/jestHelpers.js';
- *
- * // Setup the test environment
- * await setupTest(NAME, false);
- * ```
- */
-export async function setupTest(name: string, debug: boolean = false): Promise<void> {
-  expect(name).toBeDefined();
-  expect(typeof name).toBe('string');
-  expect(name.length).toBeGreaterThanOrEqual(4);
-  NAME = name;
-  HOMEDIR = path.join('.cache', 'jest', name);
-
-  // Cleanup any existing home directory
-  rmSync(HOMEDIR, { recursive: true, force: true });
-
-  const { jest } = await import('@jest/globals');
-  loggerDebugSpy = jest.spyOn(AnsiLogger.prototype, 'debug');
-  loggerInfoSpy = jest.spyOn(AnsiLogger.prototype, 'info');
-  loggerNoticeSpy = jest.spyOn(AnsiLogger.prototype, 'notice');
-  loggerWarnSpy = jest.spyOn(AnsiLogger.prototype, 'warn');
-  loggerErrorSpy = jest.spyOn(AnsiLogger.prototype, 'error');
-  loggerFatalSpy = jest.spyOn(AnsiLogger.prototype, 'fatal');
-  if (debug) {
-    loggerLogSpy = jest.spyOn(AnsiLogger.prototype, 'log');
-    consoleLogSpy = jest.spyOn(console, 'log');
-    consoleDebugSpy = jest.spyOn(console, 'debug');
-    consoleInfoSpy = jest.spyOn(console, 'info');
-    consoleWarnSpy = jest.spyOn(console, 'warn');
-    consoleErrorSpy = jest.spyOn(console, 'error');
-  } else {
-    loggerLogSpy = jest.spyOn(AnsiLogger.prototype, 'log').mockImplementation(() => {});
-    consoleLogSpy = jest.spyOn(console, 'log').mockImplementation(() => {});
-    consoleDebugSpy = jest.spyOn(console, 'debug').mockImplementation(() => {});
-    consoleInfoSpy = jest.spyOn(console, 'info').mockImplementation(() => {});
-    consoleWarnSpy = jest.spyOn(console, 'warn').mockImplementation(() => {});
-    consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
-  }
-}
-
-/**
- * Set or unset the debug mode.
- *
- * @param {boolean} debug If true, the logging is not mocked.
- * @returns {Promise<void>} A promise that resolves when the debug mode is set.
- *
- * @example
- * ```typescript
- * // Set the debug mode in test environment
- * await setDebug(true);
- * ```
- *
- * ```typescript
- * // Reset the debug mode in test environment
- * await setDebug(false);
- * ```
- */
-export async function setDebug(debug: boolean): Promise<void> {
-  const { jest } = await import('@jest/globals');
-  if (debug) {
-    loggerLogSpy.mockRestore();
-    consoleLogSpy.mockRestore();
-    consoleDebugSpy.mockRestore();
-    consoleInfoSpy.mockRestore();
-    consoleWarnSpy.mockRestore();
-    consoleErrorSpy.mockRestore();
-    loggerLogSpy = jest.spyOn(AnsiLogger.prototype, 'log');
-    consoleLogSpy = jest.spyOn(console, 'log');
-    consoleDebugSpy = jest.spyOn(console, 'debug');
-    consoleInfoSpy = jest.spyOn(console, 'info');
-    consoleWarnSpy = jest.spyOn(console, 'warn');
-    consoleErrorSpy = jest.spyOn(console, 'error');
-  } else {
-    loggerLogSpy = jest.spyOn(AnsiLogger.prototype, 'log').mockImplementation(() => {});
-    consoleLogSpy = jest.spyOn(console, 'log').mockImplementation(() => {});
-    consoleDebugSpy = jest.spyOn(console, 'debug').mockImplementation(() => {});
-    consoleInfoSpy = jest.spyOn(console, 'info').mockImplementation(() => {});
-    consoleWarnSpy = jest.spyOn(console, 'warn').mockImplementation(() => {});
-    consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
-  }
-}
 
 /**
  * Create and start a fully initialized Matterbridge instance for testing.
@@ -671,472 +555,4 @@ export async function closeMdnsInstance(matterbridge: Matterbridge): Promise<voi
   // @ts-expect-error - accessing private member for testing
   const mdns = matterbridge.environment.get(MdnsService);
   await mdns.close();
-}
-
-/**
- * Create a matter test environment for testing:
- * - it will remove any existing home directory
- * - setup the matter environment with name, debug logging and ANSI format
- * - setup the mDNS service in the environment
- *
- * @param {string} name - Name for the environment (jest/name).
- * @param {boolean} createOnly - If true, only create the environment without starting the mDNS service (default false).
- * @returns {Environment} - The default matter environment.
- */
-export function createTestEnvironment(name: string, createOnly: boolean = false): Environment {
-  expect(name).toBeDefined();
-  expect(typeof name).toBe('string');
-  expect(name.length).toBeGreaterThanOrEqual(4); // avoid accidental deletion of short paths like "/" or "C:\"
-
-  // Setup the logger
-  log = new AnsiLogger({ logName: name, logTimestampFormat: TimestampFormat.TIME_MILLIS, logLevel: LogLevel.DEBUG });
-
-  // Cleanup any existing home directory
-  rmSync(path.join(HOMEDIR), { recursive: true, force: true });
-
-  // Setup the matter environment
-  environment = Environment.default;
-  environment.vars.set('log.level', MatterLogLevel.DEBUG);
-  environment.vars.set('log.format', MatterLogFormat.ANSI);
-  environment.vars.set('path.root', path.join(HOMEDIR, '.matterbridge', MATTER_STORAGE_DIR));
-  environment.vars.set('runtime.signals', false);
-  environment.vars.set('runtime.exitcode', false);
-
-  // Return early if only creating the environment without starting the mDNS service
-  if (createOnly) return environment;
-
-  // Setup the mDNS service in the environment
-  new MdnsService(environment);
-  // await environment.get(MdnsService)?.construction.ready;
-
-  return environment;
-}
-
-/**
- * Destroy the matter test environment by closing the mDNS service.
- *
- * @param {boolean} createOnly - If true, skip destroying the environment since it was only created without starting the mDNS service (default false).
- * @returns {Promise<void>} A promise that resolves when the test environment is destroyed.
- */
-export async function destroyTestEnvironment(createOnly: boolean = false): Promise<void> {
-  if (createOnly) return;
-  // stop the mDNS service
-  const mdns = environment.get(MdnsService);
-  await mdns.close();
-}
-
-/**
- * Advance the Node.js event loop deterministically to allow chained asynchronous work (Promises scheduled in
- * microtasks and follow‑up macrotasks) to complete inside tests without adding arbitrary long timeouts.
- *
- * NOTE: This does not guarantee OS level network IO completion—only JavaScript task queue progression inside the
- *       current process.
- *
- * @param {number} ticks       Number of macrotask (setImmediate) turns to yield (default 3).
- * @param {number} microTurns  Number of microtask drains (Promise.resolve chains) after macrotask yielding (default 10).
- * @param {number} pause       Final timer delay in ms; set 0 to disable (default 250ms).
- * @returns {Promise<void>}        Resolves after the requested event loop advancement has completed.
- */
-export async function flushAsync(ticks: number = 3, microTurns: number = 10, pause: number = 250): Promise<void> {
-  for (let i = 0; i < ticks; i++) await new Promise((resolve) => setImmediate(resolve));
-  for (let i = 0; i < microTurns; i++) await Promise.resolve();
-  if (pause) await new Promise((resolve) => setTimeout(resolve, pause));
-}
-
-/**
- * Summarize live libuv handles/requests inside a process.
- *
- * @param {AnsiLogger} log - Logger to use for output
- *
- * @returns {number} - The total number of active handles and requests
- */
-export function logKeepAlives(log?: AnsiLogger): number {
-  const handles = (process as any)._getActiveHandles?.() ?? [];
-  const requests = (process as any)._getActiveRequests?.() ?? [];
-
-  // istanbul ignore next
-  const fmtHandle = (h: unknown, i: number) => {
-    const ctor = (h as { constructor?: { name?: string } })?.constructor?.name ?? 'Unknown';
-    // Timer-like?
-    const hasRef = typeof (h as any)?.hasRef === 'function' ? (h as any).hasRef() : undefined;
-    // MessagePort?
-    const isPort = (h as any)?.constructor?.name?.includes('MessagePort');
-    // Socket/Server?
-    const fd = (h as any)?.fd ?? (h as any)?._handle?.fd;
-    return { i, type: ctor, hasRef, isPort, fd };
-  };
-
-  // istanbul ignore next
-  const fmtReq = (r: unknown, i: number) => {
-    const ctor = (r as { constructor?: { name?: string } })?.constructor?.name ?? 'Unknown';
-    return { i, type: ctor };
-  };
-
-  const summary = {
-    handles: handles.map(fmtHandle),
-    requests: requests.map(fmtReq),
-  };
-
-  // istanbul ignore next if
-  if (summary.handles.length === 0 && summary.requests.length === 0) {
-    log?.debug('KeepAlive: no active handles or requests.');
-  } else {
-    log?.debug(`KeepAlive:${rs}\n${inspect(summary, { depth: 5, colors: true })}`);
-    if (!log) {
-      process.stdout.write(`KeepAlive:\n${inspect(summary, { depth: 5, colors: true })}\n`);
-    }
-  }
-  return summary.handles.length + summary.requests.length;
-}
-
-/**
- * Flush (await) the lazy endpoint number persistence mechanism used by matter.js.
- *
- * Background:
- *  assignNumber() batches persistence (store.saveNumber + updating __nextNumber__) via an internal promise (#numbersPersisted).
- *  Calling endpointStores.close() waits for the current batch only. If new endpoints were added in the same macrotask
- *  cycle additional micro/macro turns might be needed to ensure the batch started. We defensively yield macrotasks
- *  (setImmediate) and then await close() multiple rounds.
- *
- * @param {ServerNode} targetServer  The server whose endpoint numbering persistence should be flushed.
- * @param {number} rounds Number of macrotask + close cycles to run (3 is usually sufficient).
- * @param {number} pause Duration in ms to wait between cycles (default 10ms) to allow any follow-up work scheduled by close() to run.
- * @returns {Promise<void>} Resolves when pending number persistence batches have completed.
- */
-export async function flushAllEndpointNumberPersistence(targetServer: ServerNode, rounds: number = 3, pause: number = 10): Promise<void> {
-  const nodeStore = targetServer.env.get(ServerNodeStore);
-  for (let i = 0; i < rounds; i++) {
-    await flushAsync(undefined, undefined, pause);
-    await nodeStore.endpointStores.close();
-  }
-}
-
-/**
- * Get the owning ServerNode for a given Endpoint by traversing up the owner chain until the root endpoint is reached.
- *
- * @param {ServerNode<ServerNode.RootEndpoint> | Endpoint} owner The endpoint or server node to find the owning server for.
- * @returns {ServerNode<ServerNode.RootEndpoint>} The owning ServerNode of the given endpoint.
- */
-function getOwningServerNode(owner: ServerNode<ServerNode.RootEndpoint> | Endpoint): ServerNode<ServerNode.RootEndpoint> {
-  let current = owner as Endpoint;
-  while (current.owner) {
-    current = current.owner as Endpoint;
-  }
-  return current as ServerNode<ServerNode.RootEndpoint>;
-}
-
-/**
- * Collect all endpoints in the server endpoint tree (root -> descendants).
- *
- * @param {Endpoint} root  Root endpoint (typically the ServerNode root endpoint cast as Endpoint).
- * @returns {Endpoint[]}   Flat array including the root and every descendant once.
- */
-function collectAllEndpoints(root: Endpoint): Endpoint[] {
-  const list: Endpoint[] = [];
-  const walk = (ep: Endpoint) => {
-    list.push(ep);
-    if (ep.parts) {
-      for (const child of ep.parts as unknown as Endpoint[]) {
-        walk(child);
-      }
-    }
-  };
-  walk(root);
-  return list;
-}
-
-/**
- * Assert that every endpoint attached to the server has an assigned and (batch-)persisted endpoint number.
- *
- * This waits for any outstanding number persistence batch (endpointStores.close()), then traverses the endpoint
- * graph and asserts:
- *  - Root endpoint: number is 0 (allowing undefined to coerce to 0 via nullish coalescing check).
- *  - All other endpoints: number > 0.
- *
- * @param {ServerNode} targetServer The server whose endpoint numbers are verified.
- * @returns {Promise<void>}         Resolves when assertions complete.
- */
-export async function assertAllEndpointNumbersPersisted(targetServer: ServerNode): Promise<number> {
-  const nodeStore = targetServer.env.get(ServerNodeStore);
-  // Ensure any pending persistence finished (flush any in-flight batch promise)
-  await nodeStore.endpointStores.close();
-  const all = collectAllEndpoints(targetServer as unknown as Endpoint);
-  for (const ep of all) {
-    const store = nodeStore.storeForEndpoint(ep);
-    if (ep.maybeNumber === 0) {
-      expect(store.number ?? 0).toBe(0); // root
-    } else {
-      expect(store.number).toBeGreaterThan(0);
-    }
-  }
-  return all.length;
-}
-
-/**
- * Close the server node stores to flush any pending endpoint number persistence.
- *
- * @param {ServerNode} targetServer The server whose endpoint stores should be closed.
- * @returns {Promise<void>} Resolves when the stores have been closed.
- */
-export async function closeServerNodeStores(targetServer?: ServerNode): Promise<void> {
-  // Close endpoint stores to avoid number persistence issues
-  if (!targetServer) targetServer = server;
-  await targetServer?.env.get(ServerNodeStore)?.endpointStores.close();
-}
-
-/**
- * Start a matter server node for testing.
- *
- * @param {string} name Name of the server (used for logging and product description).
- * @param {number} port TCP port to listen on.
- * @param {DeviceTypeId} deviceType Device type identifier for the server node. Defaults to the bridge device type.
- * @param {number} ticks Number of macrotask (setImmediate) turns to yield after starting the server to allow asynchronous work to complete before returning (default 3).
- * @param {number} microTurns Number of microtask drains (Promise.resolve chains) to perform after macrotask yielding to allow asynchronous work to complete before returning (default 10).
- * @param {number} pause Final timer delay in ms to wait after microtask draining to allow any follow-up work scheduled by the server start process to run before returning (default 100ms).
- * @returns {Promise<[ServerNode<ServerNode.RootEndpoint>, Endpoint<AggregatorEndpoint>]>} Resolves to an array containing the created ServerNode and its AggregatorNode.
- */
-export async function createServerNode(
-  name: string,
-  port: number,
-  deviceType: DeviceTypeId = bridge.code,
-  ticks: number = 3,
-  microTurns: number = 10,
-  pause: number = 100,
-): Promise<[ServerNode<ServerNode.RootEndpoint>, Endpoint<AggregatorEndpoint>]> {
-  const { randomBytes } = await import('node:crypto');
-  const random = randomBytes(8).toString('hex');
-
-  // Create the server node
-  server = await ServerNode.create({
-    id: name + 'ServerNode',
-
-    // Provide the environment
-    environment,
-
-    // Provide Node announcement settings
-    productDescription: {
-      name: name + 'ServerNode',
-      deviceType: DeviceTypeId(deviceType),
-      vendorId: VendorId(0xfff1),
-      productId: 0x8000,
-    },
-
-    // Provide defaults for the BasicInformation cluster on the Root endpoint
-    basicInformation: {
-      vendorId: VendorId(0xfff1),
-      vendorName: 'Matterbridge',
-      productId: 0x8000,
-      productName: 'Matterbridge ' + name,
-      nodeLabel: name + 'ServerNode',
-      hardwareVersion: 1,
-      softwareVersion: 1,
-      reachable: true,
-      serialNumber: 'SN' + random,
-      uniqueId: 'UI' + random,
-    },
-
-    // Provide Network relevant configuration like the port
-    network: {
-      listeningAddressIpv4: undefined,
-      listeningAddressIpv6: undefined,
-      port,
-    },
-
-    // Provide the certificate for the device
-    operationalCredentials: {
-      certification: undefined,
-    },
-  });
-  expect(server).toBeDefined();
-  expect(server.lifecycle.isReady).toBeTruthy();
-
-  // Create the aggregator node
-  aggregator = new Endpoint(AggregatorEndpoint, {
-    id: name + 'AggregatorNode',
-  });
-  expect(aggregator).toBeDefined();
-
-  // Add the aggregator to the server
-  await server.add(aggregator);
-  expect(server.parts.has(aggregator.id)).toBeTruthy();
-  expect(server.parts.has(aggregator)).toBeTruthy();
-  expect(aggregator.lifecycle.isReady).toBeTruthy();
-
-  // Run the server
-  expect(server.lifecycle.isOnline).toBeFalsy();
-
-  // Ensure the queue is empty and pause 100ms to allow any pending work to complete before returning the server and aggregator
-  await flushAsync(ticks, microTurns, pause);
-
-  return [server, aggregator];
-}
-
-/**
- * Start a matter server node for testing.
- *
- * @param {string} name Name of the server (used for logging and product description).
- * @param {number} port TCP port to listen on.
- * @param {DeviceTypeId} deviceType Device type identifier for the server node. Defaults to the bridge device type.
- * @param {boolean} createOnly If true, only creates the server and aggregator nodes without starting the server (default false).
- * @param {number} ticks Number of macrotask (setImmediate) turns to yield after starting the server to allow asynchronous work to complete before returning (default 3).
- * @param {number} microTurns Number of microtask drains (Promise.resolve chains) to perform after macrotask yielding to allow asynchronous work to complete before returning (default 10).
- * @param {number} pause Final timer delay in ms to wait after microtask draining to allow any follow-up work scheduled by the server start process to run before returning (default 100ms).
- * @returns {Promise<[ServerNode<ServerNode.RootEndpoint>, Endpoint<AggregatorEndpoint>]>} Resolves to an array containing the created ServerNode and its AggregatorNode.
- */
-export async function startServerNode(
-  name: string,
-  port: number,
-  deviceType: DeviceTypeId = bridge.code,
-  createOnly: boolean = false,
-  ticks: number = 3,
-  microTurns: number = 10,
-  pause: number = 100,
-): Promise<[ServerNode<ServerNode.RootEndpoint>, Endpoint<AggregatorEndpoint>]> {
-  // Create the server node
-  if (!server || !aggregator) {
-    await createServerNode(name, port, deviceType);
-  }
-
-  // Return early if createOnly is true
-  if (createOnly) {
-    return [server, aggregator];
-  }
-
-  // Wait for the server to be online
-  await new Promise<void>((resolve) => {
-    server.lifecycle.online.on(async () => {
-      resolve();
-    });
-    server.start();
-  });
-
-  // Check if the server is online
-  expect(server.lifecycle.isReady).toBeTruthy();
-  expect(server.lifecycle.isOnline).toBeTruthy();
-  expect(server.lifecycle.isCommissioned).toBeFalsy();
-  expect(server.lifecycle.isPartsReady).toBeTruthy();
-  expect(server.lifecycle.hasId).toBeTruthy();
-  expect(server.lifecycle.hasNumber).toBeTruthy();
-  expect(aggregator.lifecycle.isReady).toBeTruthy();
-  expect(aggregator.lifecycle.isInstalled).toBeTruthy();
-  expect(aggregator.lifecycle.isPartsReady).toBeTruthy();
-  expect(aggregator.lifecycle.hasId).toBeTruthy();
-  expect(aggregator.lifecycle.hasNumber).toBeTruthy();
-
-  // Ensure the queue is empty and pause 100ms to allow any pending work to complete before returning the server and aggregator
-  await flushAsync(ticks, microTurns, pause);
-
-  return [server, aggregator];
-}
-
-/**
- * Stop a matter server node.
- *
- * @param {ServerNode<ServerNode.RootEndpoint>} server The server to stop.
- * @param {boolean} createOnly If true, only creates the server and aggregator nodes without starting the server (default false).
- * @param {number} ticks Number of macrotask (setImmediate) turns to yield after stopping the server to allow asynchronous work to complete before returning (default 3).
- * @param {number} microTurns Number of microtask drains (Promise.resolve chains) to perform after macrotask yielding to allow asynchronous work to complete before returning (default 10).
- * @param {number} pause Final timer delay in ms to wait after microtask draining to allow any follow-up work scheduled by the server stop process to run before returning (default 100ms).
- * @returns {Promise<void>} Resolves when the server has stopped.
- */
-export async function stopServerNode(
-  server: ServerNode<ServerNode.RootEndpoint>,
-  createOnly: boolean = false,
-  ticks: number = 3,
-  microTurns: number = 10,
-  pause: number = 100,
-): Promise<void> {
-  // Flush any pending endpoint number persistence
-  await flushAllEndpointNumberPersistence(server);
-
-  // Ensure all endpoint numbers are persisted
-  await assertAllEndpointNumbersPersisted(server);
-
-  // Stop the server
-  expect(server).toBeDefined();
-  expect(server.lifecycle.isReady).toBeTruthy();
-  if (!createOnly) {
-    expect(server.lifecycle.isOnline).toBeTruthy();
-  }
-  await server.close();
-  expect(server.lifecycle.isReady).toBeFalsy();
-  expect(server.lifecycle.isOnline).toBeFalsy();
-
-  // Ensure the queue is empty and pause 100ms to allow any pending work to complete before returning the server and aggregator
-  await flushAsync(ticks, microTurns, pause);
-}
-
-/**
- * Add a device (endpoint) to a matter server node or an aggregator.
- *
- * @param {ServerNode<ServerNode.RootEndpoint> | Endpoint<AggregatorEndpoint>} owner The server or aggregator to add the device to.
- * @param {Endpoint} device The device to add.
- * @param {number} pause The pause time in milliseconds after addition (default 10ms).
- * @returns {Promise<void>} Resolves when the device has been added and is ready.
- */
-export async function addDevice(owner: ServerNode<ServerNode.RootEndpoint> | Endpoint<AggregatorEndpoint>, device: Endpoint, pause: number = 10): Promise<boolean> {
-  expect(owner).toBeDefined();
-  expect(device).toBeDefined();
-  expect(owner.lifecycle.isReady).toBeTruthy();
-  expect(owner.construction.status).toBe(Lifecycle.Status.Active);
-  expect(owner.lifecycle.isPartsReady).toBeTruthy();
-  const targetServer = getOwningServerNode(owner);
-  await flushAllEndpointNumberPersistence(targetServer, 3, pause);
-
-  // istanbul ignore next
-  try {
-    await owner.add(device);
-  } catch (error) {
-    const errorMessage = error instanceof Error ? error.message : error;
-    const errorInspect = inspect(error, { depth: 10 });
-    process.stderr.write(`${er}Error adding device ${device.maybeId}.${device.maybeNumber}: ${errorMessage}${rs}\nStack: ${errorInspect}\n`);
-    return false;
-  }
-  await device.construction.ready;
-  expect(owner.parts.has(device)).toBeTruthy();
-  expect(owner.lifecycle.isPartsReady).toBeTruthy();
-  expect(device.lifecycle.isReady).toBeTruthy();
-  expect(device.lifecycle.isInstalled).toBeTruthy();
-  expect(device.lifecycle.hasId).toBeTruthy();
-  expect(device.lifecycle.hasNumber).toBeTruthy();
-  expect(device.construction.status).toBe(Lifecycle.Status.Active);
-  await flushAllEndpointNumberPersistence(targetServer, 3, pause);
-  return true;
-}
-
-/**
- * Delete a device (endpoint) from a matter server node or an aggregator.
- *
- * @param {ServerNode<ServerNode.RootEndpoint> | Endpoint<AggregatorEndpoint>} owner The server or aggregator to remove the device from.
- * @param {Endpoint} device The device to remove.
- * @param {number} pause The pause time in milliseconds after deletion (default 10ms).
- * @returns {Promise<void>} Resolves when the device has been removed and is no longer ready.
- */
-export async function deleteDevice(owner: ServerNode<ServerNode.RootEndpoint> | Endpoint<AggregatorEndpoint>, device: Endpoint, pause: number = 10): Promise<boolean> {
-  expect(owner).toBeDefined();
-  expect(device).toBeDefined();
-  expect(owner.lifecycle.isReady).toBeTruthy();
-  expect(owner.construction.status).toBe(Lifecycle.Status.Active);
-  expect(owner.lifecycle.isPartsReady).toBeTruthy();
-  const targetServer = getOwningServerNode(owner);
-  await flushAllEndpointNumberPersistence(targetServer, 3, pause);
-
-  // istanbul ignore next
-  try {
-    await device.delete();
-  } catch (error) {
-    const errorMessage = error instanceof Error ? error.message : error;
-    const errorInspect = inspect(error, { depth: 10 });
-    process.stderr.write(`${er}Error deleting device ${device.maybeId}.${device.maybeNumber}: ${errorMessage}${rs}\nStack: ${errorInspect}\n`);
-    return false;
-  }
-  expect(owner.parts.has(device)).toBeFalsy();
-  expect(owner.lifecycle.isPartsReady).toBeTruthy();
-  expect(device.lifecycle.isReady).toBeFalsy();
-  expect(device.lifecycle.isInstalled).toBeFalsy();
-  expect(device.lifecycle.hasId).toBeTruthy();
-  expect(device.lifecycle.hasNumber).toBeTruthy();
-  expect(device.construction.status).toBe(Lifecycle.Status.Destroyed);
-  await flushAllEndpointNumberPersistence(targetServer, 3, pause);
-  return true;
 }
